refactor: JsonView.render()
