+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JsonView.java
@@ -1,32 +1,22 @@
 package nextstep.mvc.tobe.view;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import nextstep.utils.JsonUtils;
 import nextstep.web.support.MediaType;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
+import java.io.PrintWriter;
 import java.util.Map;
 
 public class JsonView implements View {
 
     @Override
-    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception {
-        ObjectMapper objectMapper = new ObjectMapper();
+    public void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response)
+            throws Exception {
+        final PrintWriter printWriter = response.getWriter();
         setContentType(response);
-        if (model.size() == 0) {
-
-        }
-        else if (model.size() == 1) {
-            writeOnce(model, response, objectMapper);
-        }
-        else {
-            objectMapper.writeValue(response.getOutputStream(), model);
-        }
-    }
-
-    private void writeOnce(Map<String, ?> model, HttpServletResponse response, ObjectMapper objectMapper) throws IOException {
-        objectMapper.writeValue(response.getOutputStream(), model.get("car"));
+        printWriter.write(JsonUtils.toJsonString(model));
+        printWriter.flush();
     }
 
     private void setContentType(HttpServletResponse response) {
