+++ nextstep-mvc/src/test/java/nextstep/mvc/tobe/handler/HandlerExecutionTest.java
@@ -3,6 +3,8 @@ package nextstep.mvc.tobe.handler;
 import nextstep.mvc.tobe.ModelAndView;
 import nextstep.mvc.tobe.MyController;
 import org.junit.jupiter.api.Test;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
@@ -14,6 +16,7 @@ import java.lang.reflect.Method;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class HandlerExecutionTest {
+    private static final Logger logger = LoggerFactory.getLogger(HandlerExecutionTest.class);
 
     @Test
     public void handle() throws Exception {
@@ -28,5 +31,6 @@ public class HandlerExecutionTest {
         modelAndView.render(request, response);
 
         assertThat(response.getForwardedUrl()).isEqualTo("/users/view.jsp");
+        response.getOutputStream().println();
     }
 }
